requirements/dev.txt
pytest==3.0.5
pytest-describe==0.10.3
pytest-asyncio==0.5.0
asynctest==0.9.0
pyhamcrest==1.9.0
coverage==4.0.3
flake8==3.0.4
